% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function2.R
\name{ce}
\alias{ce}
\title{Confidence ellipse}
\usage{
ce(xdata, alpha)
}
\arguments{
\item{xdata}{the data frame}

\item{alpha}{the fold number of cross-validation}
}
\value{
the plot of the confidence ellipse
}
\description{
Plot the confidence ellipse of binary normal distribution data
}
\examples{
\dontrun{
    ce(ChickWeight[,1:2], 0.95)
}
}
