% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function1.R
\name{cv}
\alias{cv}
\title{A cross-validation function}
\usage{
cv(df, k = 5, seed = NULL, formula.str)
}
\arguments{
\item{df}{the data frame}

\item{k}{the fold number of cross-validation}

\item{seed}{the specific number for set.seed()}

\item{formula.str}{the formula of the linear regression model like y ~ x}
}
\value{
the estimated error of the cross-validation
}
\description{
A cross-validation function using R
}
\examples{
\dontrun{
    X <- cv(k = 5, seed = 123, formula.str = "lpsa ~ lcavol + lweight")
}
}
