% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC(N, thin, n, a, b)
}
\arguments{
\item{N}{the number of samples}

\item{thin}{the number of between-sample random numbers}

\item{n}{a parameter of the binomial distribution}

\item{a}{a parameter of the beta df}

\item{b}{a parameter of the beta df}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
  rnC <- gibbsC(100,10)
}
}
