\name{StatComp21037-package}
\alias{StatComp21037-package}
\alias{StatComp21037}
\docType{package}
\title{
  The final homework of 'Statistical Computing' course
}
\description{
  Several R, Rcpp function, homework and two vignettes files are included in this package for illustration.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Ruiqi Chen

Maintainer: Ruiqi Chen <accelerator@mail.ustc.edu.cn>
}
\keyword{ package }
